refactor(frontend): extract NavItem into its own component file
